Added Audio Engine
Audio file loading(Vorbis OGG loading is crashing and WAV file loading
not implemented)

+++ src/com/base/RenderingEngine/mesh/Circle.java
@@ -10,18 +10,20 @@ import org.joml.Vector2f;
  */
 public class Circle extends Mesh
 {
-	
+
 	/** The radius. */
 	private float radius;
-	
+
 	/** The resolution. */
 	private int resolution;
-	
+
 	/**
 	 * Instantiates a new circle.
 	 *
-	 * @param radius the radius
-	 * @param resolution the resolution of the triangles
+	 * @param radius
+	 *            the radius
+	 * @param resolution
+	 *            the resolution of the triangles
 	 */
 	public Circle(float radius, int resolution)
 	{
@@ -29,42 +31,44 @@ public class Circle extends Mesh
 		this.resolution = resolution;
 		this.init();
 	}
-	
-	/* (non-Javadoc)
+
+	/*
+	 * (non-Javadoc)
+	 * 
 	 * @see com.base.RenderingEngine.mesh.Mesh#initVertices()
 	 */
 	@Override
-	protected void initVertices() 
+	protected void initVertices()
 	{
 		Vertex[] cVertex = new Vertex[resolution + 1];
-		cVertex[0] = new Vertex(new Vector2f(0,0), new Vector2f(.5f,.5f));
-		float angle = 360.0f/(float)resolution;
-		
+		cVertex[0] = new Vertex(new Vector2f(0, 0), new Vector2f(.5f, .5f));
+		float angle = 360.0f / resolution;
+
 		for(int i = 0; i < resolution; i++)
 		{
-			float a = i*angle;
-			float cosX = (float)Math.cos(a);
-			float sinY = (float)Math.sin(a);
-			
-			cVertex[i+1] = new Vertex(new Vector2f(cosX, sinY).mul(radius), new Vector2f((cosX+1)/2, (sinY+1)/2));
+			float a = i * angle;
+			float cosX = (float) Math.cos(a);
+			float sinY = (float) Math.sin(a);
+
+			cVertex[i + 1] = new Vertex(new Vector2f(cosX, sinY).mul(radius), new Vector2f((cosX + 1) / 2, (sinY + 1) / 2));
 		}
-		
+
 		Stack<Integer> indStack = new Stack<>();
-		for(int i = 2; i < ((3*resolution)-1); i++)
+		for(int i = 2; i < ((3 * resolution) - 1); i++)
 		{
 			indStack.push(0);
-			indStack.push(i-1);
+			indStack.push(i - 1);
 			indStack.push(i);
 		}
-		
+
 		Object[] inds = indStack.toArray();
-		
+
 		int[] indices = new int[inds.length];
 		for(int i = 0; i < inds.length; i++)
 		{
-			indices[i] = (Integer)inds[i];
+			indices[i] = (Integer) inds[i];
 		}
-		
+
 		this.bufferVertices(cVertex, indices);
 	}
